requirements.txt
anthropic==0.34.2
PyPDF2==3.0.1
pillow==10.3.0
pdf2image==1.17.0
python-dotenv==1.0.1